# Add new `Performance/StringIdentifierArgument` cop