--- a/changelog/new_add_new_performance_string_identifier_argument_cop.md
+++ b/changelog/new_add_new_performance_string_identifier_argument_cop.md
@@ -1,0 +1,1 @@
+* [#276](https://github.com/rubocop/rubocop-performance/pull/276): Add new `Performance/StringIdentifierArgument` cop. ([@koic][])