This PR adds new `Performance/StringIdentifierArgument` cop.

This cop identifies places where string identifier argument can be replaced by symbol identifier argument. It prevents the redundancy of the internal string-to-symbol conversion.

This cop targets methods that take identifier (e.g. method name) argument and the following examples are parts of it.

```ruby
# bad
send('do_something')
attr_accessor 'do_something'
instance_variable_get('@ivar')

# good
send(:do_something)
attr_accessor :do_something
instance_variable_get(:@ivar)
```

The `send` method is a representative and other similar methods are similar benchmarks. The following is an example benchmark.

```ruby
% cat symbol.rb
puts `ruby -v`

require 'benchmark/ips'

def foo
end

Benchmark.ips do |x|
  x.report('symbol arg') { send(:foo) }
  x.report('string arg') { send('foo') }

  x.compare!
end
```

```console
% ruby symbol.rb
ruby 3.1.0dev (2021-12-05T10:23:42Z master 19f037e452) [x86_64-darwin19]
Warming up --------------------------------------
          symbol arg     1.025M i/100ms
          string arg   590.038k i/100ms
Calculating -------------------------------------
          symbol arg     10.665M (± 1.5%) i/s -     53.318M in 5.000551s

          string arg      5.895M (± 1.0%) i/s -     29.502M in 5.004999s

Comparison:
          symbol arg: 10665035.8 i/s
          string arg:  5895132.3 i/s - 1.81x  (± 0.00) slower
```

I learned about this performance difference from the book "Polished Ruby Programming".

-----------------

Before submitting the PR make sure the following are checked:

* [x] The PR relates to *only* one subject with a clear title and description in grammatically correct, complete sentences.
* [x] Wrote [good commit messages][1].
* [ ] Commit message starts with `[Fix #issue-number]` (if the related issue exists).
* [x] Feature branch is up-to-date with `master` (if not - rebase it).
* [x] Squashed related commits together.
* [x] Added tests.
* [x] Ran `bundle exec rake default`. It executes all tests and runs RuboCop on its own code.
* [x] Added an entry (file) to the [changelog folder](https://github.com/rubocop/rubocop-performance/blob/master/changelog/) named `{change_type}_{change_description}.md` if the new code introduces user-observable changes. See [changelog entry format](https://github.com/rubocop/rubocop/blob/master/CONTRIBUTING.md#changelog-entry-format) for details.

[1]: https://chris.beams.io/posts/git-commit/